refactor: Item 엔티티의 price 자료형 변경(Long -> int형)

+++ src/main/java/com/programmers/shoppingMall/entity/Item.java
@@ -31,7 +31,7 @@ public class Item {
     private String itemName; //상품명
 
     @Column(name = "price",nullable = false)
-    private Long price;//가격
+    private int price;//가격
 
     @Lob
     @Column(nullable = false)
